requirements.txt
Flask==2.0.2
fastapi==0.78.0
uvicorn==0.18.2
transformers==4.25.1
torch==2.6.0
sentence-transformers==2.2.2
scikit-learn
pandas
numpy<2
scikit-surprise